+++ p5-server/tests/testdata/syntax-error/syntax-error.js
@@ -1,0 +1,7 @@
+function setup() {
+    createCanvas(windowWidth, windowHeight);
+}
+
+function draw() {
+    rect(0, 0, 100, 100;
+}